refactor(ava/advisor): init new advisor pipeline & built-in modules to plugin-based architecture
