--- packages/ava/src/advisor/advise-pipeline/chart-type-recommend/index.ts
@@ -1,1 +1,0 @@
-export * from './get-chart-Type';
