--- packages/ava/src/advisor/advise-pipeline/data-processors/index.ts
@@ -1,2 +1,0 @@
-export * from './get-data-properties';
-export * from './get-selected-data';
