+++ packages/ava/src/advisor/advise-pipeline/data-to-advices.ts
@@ -1,16 +1,17 @@
 import { cloneDeep, deepMix } from '../utils';
 
-import { getChartTypeSpec } from './spec-generator';
+import { getChartTypeSpec } from './plugin/presets/spec-generator';
 import { DEFAULT_COLOR } from './constants';
-import { getChartTypeRecommendations } from './chart-type-recommend/get-chart-Type';
-import { applyTheme, applyDesignRules, applySmartColor } from './spec-processors';
-import { getDataProps, getSelectedData } from './data-processors';
+import { getChartTypeRecommendations } from './plugin/presets/chart-type-recommend/get-chart-Type';
+import { applyTheme, applyDesignRules, applySmartColor } from './plugin/presets/spec-generator/spec-processors';
+import { getDataProps, getSelectedData } from './plugin/presets/data-processors';
 
 import type { ScoringResultForChartType, Advice, AdviseResult, ChartAdviseParams } from '../types';
 import type { RuleModule } from '../ruler/types';
 import type { ChartKnowledgeBase } from '../../ckb';
 
 /**
+ * @deprecated 已改造为 plugin 插件形式，之前的硬编码形式函数后续清理掉
  * recommending charts given data and dataProps, based on CKB and RuleBase
  *
  * @param params input params for charts recommending
