+++ packages/ava/src/advisor/advise-pipeline/index.ts
@@ -1,3 +1,3 @@
 export * from './data-to-advices';
-export * from './spec-generator';
-export * from './data-processors';
+export * from './plugin/presets/spec-generator';
+export * from './plugin/presets/data-processors';
