+++ packages/ava/src/advisor/advise-pipeline/plugin/presets/chart-type-recommend/index.ts
@@ -1,0 +1,2 @@
+export * from './get-chart-Type';
+export { chartTypeRecommendPlugin } from './plugin-config';
