+++ packages/ava/src/advisor/advise-pipeline/plugin/presets/data-processors/index.ts
@@ -1,0 +1,3 @@
+export * from './get-data-properties';
+export * from './get-selected-data';
+export { dataProcessorPlugin } from './plugin-config';
