+++ packages/ava/src/advisor/advise-pipeline/plugin/presets/spec-generator/index.ts
@@ -1,0 +1,2 @@
+export { getChartTypeSpec } from './get-chart-spec';
+export { specGeneratorPlugin } from './plugin-config';
