+++ packages/ava/src/advisor/advise-pipeline/plugin/presets/visual-encoder/index.ts
@@ -1,0 +1,1 @@
+export { visualEncoderPlugin } from './plugin-config';
